Moved version check to utils.py
- added some notes
- changed Dev version to 1.7.7
- Changed ManuelBastioni_Character to MB_LAB_Character
