+++ morphengine.py
@@ -29,7 +29,7 @@ import bpy
 
 import mathutils
 
-from . import algorithms, proxyengine
+from . import algorithms, utils, proxyengine
 
 import time, json
 import operator
