+++ proxyengine.py
@@ -29,7 +29,7 @@ import os
 import mathutils
 import bpy
 
-from . import algorithms
+from . import algorithms, utils
 
 
 logger = logging.getLogger(__name__)
